Added docs skeleton v2, removed solution helper
