--- business-logic/solution-helper/package.json
@@ -1,23 +1,0 @@
-{
-    "name": "game-analytics-solution-helper",
-    "description": "Game Analytics Pipeline solution custom resource helper Lambda function",
-    "main": "index.js",
-    "version": "0.0.1",
-    "private": true,
-    "dependencies": {
-        "aws-sdk": "*",
-        "moment": "*",
-        "underscore": "*",
-        "uuid": "*"
-    },
-    "devDependencies": {
-        "aws-sdk": "*",
-        "npm-run-all": "*"
-    },
-    "scripts": {
-        "build:init": "rm -rf package-lock.json && rm -rf dist && rm -rf node_modules",
-        "build:zip": "zip -rq solution-helper.zip .",
-        "build:dist": "mkdir dist && mv solution-helper.zip dist/",
-        "build": "npm run build:init && npm install --production && npm run build:zip && npm run build:dist"
-    }
-}
