+++ docs/component-deep-dive.md
@@ -1,0 +1,14 @@
+# Component Deep Dive
+
+![Architecture](media/architecture.png)
+
+## Part 1. Ingest
+Sending events to the Game Analytics Pipeline are through REST HTTPS API calls. 
+
+## Part 2. Store
+
+## Part 3. Process
+
+## Part 4. Visualize
+
+## Part 5. Manage
