+++ docs/design-considerations.md
@@ -1,0 +1,8 @@
+# Design Considerations
+
+- If there are any differing opinions, encourage discussion through github issues
+
+- Why only REST HTTP API, why not send passthrough or direct?
+- Why extra lambdas?
+- Why DynamoDB with application_id 
+- why is the Schema the way we designed it?
