+++ docs/further-tuning.md
@@ -1,0 +1,13 @@
+# Further Tuning
+
+## Custom events + queries (sample walkthroughs)
+    - non-real-time
+    - real-time
+
+## Custom ETL
+
+## Modifying schema
+
+## Modifying/extending architecture
+
+## (TODO: Send from game engines)
