+++ docs/security.md
@@ -1,0 +1,14 @@
+---
+title: Security
+description: Security practices and processes for Game Analytics Pipeline Guidance
+---
+
+## Overview
+
+This project is maintained by members of the AWS for Games technical community within AWS (i.e. Solutions Architects, Technical Account Managers, Software Engineers) who support the gaming industry. Design decisions and tradeoffs made throughout this project reflect our experiences working with game studios to build and maintain their development infrastructure and tools in the cloud. We encourage contributions from the community via Pull Requests, which are manually reviewed and tested by the core maintainers of the project before they are merged.
+
+## Reporting a vulnerability
+
+If you discover a potential security issue in this project, we ask that you notify AWS/Amazon Security via our [vulnerability reporting page](http://aws.amazon.com/security/vulnerability-reporting/) or directly via email to <aws-security@amazon.com>.
+
+Please do **not** create a public GitHub issue.
