--- docs/upgrading/v1-migration.md
@@ -1,1 +1,0 @@
-# Upgrading from Game Analytics Pipeline on AWS Version 1
