+++ mkdocs.yml
@@ -1,13 +1,17 @@
-site_name: Game Analytics Pipeline on AWS
+site_name: Guidance for Game Analytics Pipeline on AWS
 site_description: This Guidance helps game developers deploy a modular, serverless analytics pipeline for ingesting, storing, processing, and visualizing telemetry data from game clients and backend services. It supports real-time and batch analytics use cases. Using AWS Cloud Development Kit, developers can continuously integrate and deploy the pipeline across multiple AWS accounts and Regions. The serverless services offer cost-efficiency and ease of setup. After deployment, teams can ingest data and make informed decisions to enhance their games faster.
 site_author: Amazon Web Services
 repo_url: https://github.com/aws-solutions-library-samples/guidance-for-game-analytics-pipeline-on-aws
 repo_name: aws-solutions-library-samples/guidance-for-game-analytics-pipeline-on-aws
+site_url: https://aws.amazon.com/solutions/guidance/game-analytics-pipeline-on-aws/
+edit_uri: edit/main/docs/
 
 theme:
   name: material
+  custom_dir: docs/overrides
   favicon: media/aws.png
   font: false
+  logo: media/aws-logo-white.svg
   palette:
     # Palette toggle for automatic mode
     - media: "(prefers-color-scheme)"
@@ -30,21 +34,85 @@ theme:
       toggle:
         icon: material/lightbulb-outline
         name: Switch to light mode
+
   features:
+    - toc:
+        permalink: true
+        sticky: true
+        integrate: true
+    - content.code.annotate
+    - content.code.copy
+    - content.tabs.link
+    - search.suggest
+    - search.highlight
+    - search.share
+    - navigation.sections
+    - navigation.indexes
+    - navigation.tracking
     - navigation.tabs
     - navigation.tabs.sticky
+    - navigation.top
     - navigation.expand
+    - navigation.instant
+    - navigation.footer
+    - toc.integrate
+    - emoji_index: twemoji
+
+markdown_extensions:
+  - meta
+  - admonition
+  - footnotes
+  - pymdownx.details
+  - pymdownx.highlight
+  - pymdownx.superfences
+  - pymdownx.inlinehilite
+  - attr_list
+  - md_in_html
+  - def_list
+  - tables
+  - pymdownx.snippets:
+      base_path: "."
+      check_paths: true
+      restrict_base_path: false
+  - toc:
+      permalink: true
+      toc_depth: 5
+  - pymdownx.emoji:
+      emoji_index: !!python/name:material.extensions.emoji.twemoji
+      emoji_generator: !!python/name:material.extensions.emoji.to_svg
+  - pymdownx.tabbed:
+      alternate_style: true
+
 
 extra:
-  homepage: https://aws-solutions-library-samples.github.io/
+  homepage: index.html
 
 nav:
   - Home: index.md
   - Getting Started: getting-started.md
   - Upgrading:
-      - Upgrading from Game Analytics Pipeline on AWS Version 1: upgrading/v1-migration.md
-      - Upgrading from Kinesis Data Analytics to Managed Apache Flink: upgrading/flink-migration.md
-  - See this use case in AWS Solutions Library: https://aws.amazon.com/solutions/guidance/game-analytics-pipeline-on-aws/
+      - Changelog: upgrading/v2-to-v3-changes.md
+      - Upgrade Guide: upgrading/v2-migration.md
+      - (old-flink-remove): upgrading/flink-migration.md
+  - Component Deep Dive: component-deep-dive.md
+  - Further Tuning: further-tuning.md
+  - References:
+    - API Reference: references/api-reference.md
+    - Schema Reference: references/schema-reference.md
+    - Operational Dashboard Reference: references/ops-dashboard-reference.md
+  - Design Considerations: design-considerations.md
+  - Roadmap: roadmap.md
+  - Security: security.md
+  - Contributing: contributing.md
+  - AWS Landing Page: https://aws.amazon.com/solutions/guidance/game-analytics-pipeline-on-aws/
+
+extra_css:
+  - stylesheets/extra.css
+  - stylesheets/themes.css
+  - stylesheets/hero.css
+
 plugins:
   - search
   - offline
+
+copyright: Copyright &copy; 2024 Amazon Web Services
